// constants/enum.ts
export enum EImportingClient {
  PING_OK,
  FAILED = 'FAILED',
  CLIENT_CREATED = 'CLIENT_CREATED',
  IMPORTED_CHECKPOINT = 'IMPORTED_CHECKPOINT',
  IMPORTED_LORA = 'IMPORTED_LORA',
  IMPORTED_SAMPLER_SCHEDULER = 'IMPORTED_SAMPLER_SCHEDULER',
  IMPORTED_EXTENSION = 'IMPORTED_EXTENSION',
  DONE = 'DONE'
}